# internal/model/model.go
package model

type Type string

const (
	NORMAL  Type = "normal"
	EXPRESS Type = "express"
)

type Message struct {
	CustomerID        int64    `json:"customer_id"`
	Text              string   `json:"text"`
	Recipients        []string `json:"recipients"`
	Type              Type     `json:"type"`
	TransactionID     string   `json:"transaction_id"`
	MessageIdentifier string   `json:"message_identifier"`
}